requirements.txt
cloudinary
Django
djangorestframework
djangorestframework-simplejwt
django-filter
librosa
Pillow
pydub
requests
django-cloudinary-storage
mutagen
python-decouple